Properly balance keys on layers 0 and 1
Rotate a few keys around in order to balance thumb movement for layer 0.
The previous config had essentialy 4 buttons on the left thumb, and only
2 on the right. Move to an equal 3/3.

Eliminate "trans" bindings on layer 1, in order to reduce the factor of
suprise that appeared while holding the layer key and pressing random
buttons that were not supposed to be mapped on layer 1.