Disable RGB underglow, fix battery notification bug
Disable the RGB underglow, in order to be able to use the OLED display,
while the RGB's are not pulling current while still idle (I am still not
sure if disabling the RGB strip in the config completely cuts off
external power to the RGB LED's, however monitoring the battery usage
should give me a rough idea).

Fix notification bug in windows which prevents the battery of the
keyboard from being reported.