Change to 10 second delay on changes save
The default on saving changes to flash memory is 60 seconds. This saves
things like bluetooth profiles, external power settings and so on.
Reduce this to 10 second only for the external power switch, since the
current ZMK bug still prevents an OLED power up after external power is
switched on, and the only way to re-start it properly is to wait the 60
second delay, and reboot the keyboard.